+++ Include/Goblin/User/Interface/Announce.hpp
@@ -22,7 +22,7 @@ along with Goblin Camp. If not, see <http://www.gnu.org/licenses/>.*/
 #include <libtcod.hpp>
 
 #include "Goblin/Geometry/Coordinate.hpp"
-#include "Goblin/User/Interface/UIComponents.hpp"
+#include "Goblin/User/Interface/Drawable.hpp"
 
 #define ANNOUNCE_MAX_LENGTH 71
 #define ANNOUNCE_HEIGHT 10
