+++ Include/Goblin/User/Interface/Frame.hpp
@@ -19,7 +19,7 @@
 #include <vector>
 #include <libtcod.hpp>
 
-#include "Goblin/User/Interface/UIComponents.hpp"
+#include "Goblin/User/Interface/Drawable.hpp"
 
 class Frame : public UIContainer
 {
