+++ Include/Goblin/User/Interface/JobDialog.hpp
@@ -17,7 +17,7 @@
 
 #include <libtcod.hpp>
 
-#include "Goblin/User/Interface/UIComponents.hpp"
+#include "Goblin/User/Interface/Drawable.hpp"
 #include "Goblin/User/Interface/Dialog.hpp"
 
 class JobDialog : public Scrollable
