+++ Include/Goblin/User/Interface/Spinner.hpp
@@ -18,7 +18,7 @@
 #include <functional>
 #include <libtcod.hpp>
 
-#include "Goblin/User/Interface/UIComponents.hpp"
+#include "Goblin/User/Interface/Drawable.hpp"
 
 class Spinner : public Drawable
 {
